+++ build.gradle
@@ -24,7 +24,6 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-security'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'io.springfox:springfox-boot-starter:3.0.0'
-	implementation 'org.springdoc:springdoc-openapi-ui:1.6.9'
 	implementation 'mysql:mysql-connector-java:8.0.30'
 	implementation 'org.mockito:mockito-inline:4.11.0'
 	implementation 'com.vladmihalcea:hibernate-types-52:2.17.3'
