--- src/main/java/com/postype/sns/application/contoller/HealthCheckController.java
@@ -1,13 +1,0 @@
-package com.postype.sns.application.contoller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthCheckController {
-
-	@GetMapping("/")
-	public String healthCheck(){
-		return "The service is up and running...";
-	}
-}
