+++ src/main/java/com/postype/sns/application/controller/FollowController.java
@@ -1,18 +1,15 @@
-package com.postype.sns.application.contoller;
+package com.postype.sns.application.controller;
 
-import com.postype.sns.application.contoller.dto.response.FollowResponse;
-import com.postype.sns.application.contoller.dto.response.Response;
-import com.postype.sns.application.contoller.dto.FollowDto;
-import com.postype.sns.application.contoller.dto.MemberDto;
-import com.postype.sns.application.exception.ApplicationException;
-import com.postype.sns.application.exception.ErrorCode;
+import com.postype.sns.application.controller.dto.response.FollowResponse;
+import com.postype.sns.application.controller.dto.response.Response;
+import com.postype.sns.application.controller.dto.FollowDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.service.FollowService;
 import com.postype.sns.domain.member.service.MemberService;
-import com.postype.sns.utill.ClassUtils;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,6 +25,7 @@ public class FollowController {
 	private final FollowService followService;
 	private final MemberService memberService;
 
+	@Operation(summary = "팔로우", description = "로그인한 멤버가 toMemberName을 팔로우 합니다")
 	@PostMapping("/{toMemberName}")
 	public Response<FollowResponse> create(@AuthenticationPrincipal MemberDto fromMember, @PathVariable String toMemberName){
 		MemberDto toMember = memberService.getMember(toMemberName);
@@ -36,6 +34,7 @@ public class FollowController {
 	}
 
 	//fromId가 팔로잉 하고 있는 목록 확인할 수 있음
+	@Operation(summary = "팔로잉 리스트 가져오기", description = "로그인한 멤버의 팔로잉 리스트를 가져옵니다")
 	@GetMapping
 	public Response<Page<FollowResponse>> getFollowList(@AuthenticationPrincipal MemberDto memberDto, Pageable pageable){
 		return Response.success(followService.getFollowList(memberDto, pageable).map(FollowResponse::fromFollowDto));
