+++ src/main/java/com/postype/sns/application/controller/OrderController.java
@@ -1,26 +1,18 @@
-package com.postype.sns.application.contoller;
+package com.postype.sns.application.controller;
 
-import com.postype.sns.application.contoller.dto.MemberDto;
-import com.postype.sns.application.contoller.dto.request.OrderCreateRequest;
-import com.postype.sns.application.contoller.dto.response.OrderResponse;
-import com.postype.sns.application.contoller.dto.response.Response;
-import com.postype.sns.application.exception.ApplicationException;
-import com.postype.sns.application.exception.ErrorCode;
+import com.postype.sns.application.controller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.response.OrderResponse;
+import com.postype.sns.application.controller.dto.response.Response;
 import com.postype.sns.application.usecase.OrderUseCase;
-import com.postype.sns.domain.order.model.Order;
-import com.postype.sns.utill.ClassUtils;
-import java.util.List;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -29,17 +21,19 @@ import org.springframework.web.bind.annotation.RestController;
 public class OrderController {
 
 	private final OrderUseCase OrderUseCase;
-
+	@Operation(summary = "포스트 구입", description = "로그인한 멤버가 postId에 해당하는 포스트를 구입합니다")
 	@PostMapping("/{postId}")
 	public Response<OrderResponse> create(@AuthenticationPrincipal MemberDto memberDto, @PathVariable Long postId){
 		return Response.success(
 			OrderResponse.fromDto(OrderUseCase.create(memberDto, postId)));
 	}
+	@Operation(summary = "포스트 구입 내역 확인하기", description = "로그인한 멤버가 구입한 postId의 구입내역을 가져옵니다")
 	@GetMapping("/{postId}")
 	public Response<OrderResponse> getOrderByMemberAndPost(@AuthenticationPrincipal MemberDto memberDto, @PathVariable Long postId){
 		return Response.success(
 			OrderResponse.fromDto(OrderUseCase.getOrderByMemberAndPost(memberDto, postId)));
 	}
+	@Operation(summary = "멤버의 구입 내역 가져오기", description = "로그인한 멤버의 모든 구입내역을 가져옵니다")
 	@GetMapping
 	public Response<Page<OrderResponse>> getOrder(@AuthenticationPrincipal MemberDto memberDto, Pageable pageable){
 		return Response.success(
