+++ src/main/java/com/postype/sns/application/controller/dto/FollowDto.java
@@ -1,4 +1,4 @@
-package com.postype.sns.application.contoller.dto;
+package com.postype.sns.application.controller.dto;
 
 import com.postype.sns.domain.member.model.Follow;
 import java.sql.Timestamp;
