+++ src/main/java/com/postype/sns/application/controller/dto/MemberDto.java
@@ -1,4 +1,4 @@
-package com.postype.sns.application.contoller.dto;
+package com.postype.sns.application.controller.dto;
 
 import com.postype.sns.domain.member.model.Member;
 import com.postype.sns.domain.member.model.MemberRole;
