+++ src/main/java/com/postype/sns/application/controller/dto/OrderDto.java
@@ -1,4 +1,4 @@
-package com.postype.sns.application.contoller.dto;
+package com.postype.sns.application.controller.dto;
 
 
 import com.postype.sns.domain.order.model.Order;
