+++ src/main/java/com/postype/sns/application/controller/dto/request/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package com.postype.sns.application.contoller.dto.request;
+package com.postype.sns.application.controller.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,10 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class MemberLoginRequest{
+public class MemberRegisterRequest{
+
 	private String memberId;
 	private String password;
+	private String memberName;
+	private String email;
 }
