+++ src/main/java/com/postype/sns/application/controller/dto/request/PostModifyRequest.java
@@ -1,13 +1,12 @@
-package com.postype.sns.application.contoller.dto.request;
+package com.postype.sns.application.controller.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class PostCreateRequest {
+public class PostModifyRequest {
 	private String title;
 	private String body;
 	private int price;
