+++ src/main/java/com/postype/sns/application/controller/dto/response/AlarmResponse.java
@@ -1,6 +1,6 @@
-package com.postype.sns.application.contoller.dto;
+package com.postype.sns.application.controller.dto.response;
 
-import com.postype.sns.domain.member.model.Alarm;
+import com.postype.sns.application.controller.dto.AlarmDto;
 import com.postype.sns.domain.member.model.AlarmArgs;
 import com.postype.sns.domain.member.model.AlarmType;
 import java.sql.Timestamp;
@@ -9,20 +9,22 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class AlarmDto {
+public class AlarmResponse {
 	private Long id;
 	private AlarmType alarmType;
 	private AlarmArgs alarmArgs;
+	private String message;
 	private Timestamp registerAt;
 	private Timestamp updatedAt;
 	private Timestamp deletedAt;
 
-	public static AlarmDto fromEntity(Alarm alarm){
-		return new AlarmDto(
+	public static AlarmResponse fromDto(AlarmDto alarm){
+		return new AlarmResponse(
 			alarm.getId(),
 			alarm.getAlarmType(),
 			alarm.getAlarmArgs(),
-			alarm.getRegisteredAt(),
+			alarm.getAlarmType().getMessage(),
+			alarm.getRegisterAt(),
 			alarm.getUpdatedAt(),
 			alarm.getDeletedAt()
 		);
