+++ src/main/java/com/postype/sns/application/controller/dto/response/FollowResponse.java
@@ -1,6 +1,6 @@
-package com.postype.sns.application.contoller.dto.response;
+package com.postype.sns.application.controller.dto.response;
 
-import com.postype.sns.application.contoller.dto.FollowDto;
+import com.postype.sns.application.controller.dto.FollowDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
