+++ src/main/java/com/postype/sns/application/controller/dto/response/MemberRegisterResponse.java
@@ -1,6 +1,6 @@
-package com.postype.sns.application.contoller.dto.response;
+package com.postype.sns.application.controller.dto.response;
 
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.model.MemberRole;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
