+++ src/main/java/com/postype/sns/application/exception/GlobalControllerAdvice.java
@@ -1,8 +1,7 @@
 package com.postype.sns.application.exception;
 
-import com.postype.sns.application.contoller.dto.response.Response;
+import com.postype.sns.application.controller.dto.response.Response;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
