+++ src/main/java/com/postype/sns/application/usecase/PostUseCase.java
@@ -1,9 +1,8 @@
 package com.postype.sns.application.usecase;
 
-import com.postype.sns.application.contoller.dto.FollowDto;
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.FollowDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.service.FollowService;
-import com.postype.sns.domain.member.service.MemberService;
 import com.postype.sns.domain.post.service.PostService;
 import com.postype.sns.domain.post.service.TimeLineService;
 import java.util.List;
