+++ src/main/java/com/postype/sns/domain/order/repository/OrderRepository.java
@@ -1,15 +1,12 @@
 package com.postype.sns.domain.order.repository;
 
-import com.postype.sns.application.contoller.dto.MemberDto;
-import com.postype.sns.application.contoller.dto.OrderDto;
+import com.postype.sns.application.controller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.OrderDto;
 import com.postype.sns.domain.order.model.Order;
-import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
