+++ src/main/java/com/postype/sns/domain/order/service/OrderService.java
@@ -2,13 +2,13 @@ package com.postype.sns.domain.order.service;
 
 import com.postype.sns.application.exception.ApplicationException;
 import com.postype.sns.application.exception.ErrorCode;
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.model.Member;
 import com.postype.sns.domain.order.model.Order;
-import com.postype.sns.application.contoller.dto.OrderDto;
+import com.postype.sns.application.controller.dto.OrderDto;
 import com.postype.sns.domain.order.repository.OrderRepository;
 import com.postype.sns.domain.post.model.Post;
-import com.postype.sns.application.contoller.dto.PostDto;
+import com.postype.sns.application.controller.dto.PostDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
