+++ src/main/java/com/postype/sns/domain/post/model/Post.java
@@ -1,11 +1,9 @@
 package com.postype.sns.domain.post.model;
 
-import com.postype.sns.application.contoller.dto.PostDto;
+import com.postype.sns.application.controller.dto.PostDto;
 import com.postype.sns.domain.BaseDateEntity;
 import com.postype.sns.domain.member.model.Member;
 import com.postype.sns.domain.order.model.Point;
-import java.sql.Timestamp;
-import java.time.Instant;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -14,8 +12,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
 import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
