+++ src/test/java/com/postype/sns/application/controller/FollowControllerTest.java
@@ -9,8 +9,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.postype.sns.application.exception.ApplicationException;
 import com.postype.sns.application.exception.ErrorCode;
-import com.postype.sns.application.contoller.dto.FollowDto;
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.FollowDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.service.FollowService;
 import com.postype.sns.domain.member.service.MemberService;
 import org.junit.jupiter.api.DisplayName;
