+++ src/test/java/com/postype/sns/domain/member/service/FollowServiceTest.java
@@ -6,7 +6,7 @@ import static org.mockito.Mockito.when;
 
 import com.postype.sns.application.exception.ApplicationException;
 import com.postype.sns.application.exception.ErrorCode;
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.model.Follow;
 import com.postype.sns.domain.member.model.Member;
 import com.postype.sns.domain.member.repository.FollowRepository;
