+++ src/test/java/com/postype/sns/domain/member/service/OrderServiceTest.java
@@ -6,14 +6,14 @@ import static org.mockito.Mockito.when;
 
 import com.postype.sns.application.exception.ApplicationException;
 import com.postype.sns.application.exception.ErrorCode;
-import com.postype.sns.application.contoller.dto.MemberDto;
+import com.postype.sns.application.controller.dto.MemberDto;
 import com.postype.sns.domain.member.model.Member;
 import com.postype.sns.domain.member.repository.MemberRepository;
 import com.postype.sns.domain.order.model.Order;
 import com.postype.sns.domain.order.repository.OrderRepository;
 import com.postype.sns.domain.order.service.OrderService;
 import com.postype.sns.domain.post.model.Post;
-import com.postype.sns.application.contoller.dto.PostDto;
+import com.postype.sns.application.controller.dto.PostDto;
 import com.postype.sns.domain.post.repository.PostRepository;
 import com.postype.sns.fixture.MemberFixture;
 import com.postype.sns.fixture.OrderFixture;
